* Trying to determine where concurrent modifications exist

+++ trunk/components/component.host.js
@@ -75,8 +75,9 @@ var HostComponent = LogicComponent.extend(/** @scope HostComponent.prototype */{
     * @private
     */
    destroy: function() {
-      this.objects.destroy();
-      this.base();
+      if (this.base()) {
+         this.objects.destroy();
+      }
    },
 
    /**
